test(lsp): move tests out of source files
