+++ crates/lsp/src/tests/completion.rs
