+++ crates/lsp/src/tests/goto_definition.rs
