+++ crates/lsp/src/tests/inlay_hint.rs
