+++ crates/lsp/src/tests/references.rs
