+++ crates/lsp/src/tests/support.rs
